Place authentication in importable file for use with test cases placed in multiple files
